pyproject.toml
[project]
name = "agromind"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    # LangChain + Vertex AI
    "langchain>=0.3",
    "langchain-google-vertexai>=2.0",
    "langchain-community>=0.3",
    "langchain-chroma>=0.2",
    "chromadb>=0.5",

    # Web framework
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "httpx>=0.27",

    # Config
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",

    # Firebase
    "firebase-admin>=6.5",

    # Document AI (OCR for scanned PDFs)
    "google-cloud-documentai>=3.0",
    "langchain-google-community>=3.0",

    # Data processing
    "pandas>=2.0",
    "pypdf>=4.0",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",

    # Voice
    # sarvam-ai SDK or raw httpx calls — add when Phase 8 begins
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.5",
    "mypy>=1.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agromind"]

[tool.ruff]
line-length = 99
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
